# outputs/openapi.yaml
openapi: 3.1.0
info:
  title: Home Cleaning Planner API
  version: 0.1.0
  description: REST API contract for the cleaning planner MVP.
servers:
  - url: https://api.example.com/v1
    description: Production placeholder
  - url: http://localhost:3000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /households/{householdId}/members:
    get:
      tags: [Members]
      operationId: listHouseholdMembers
      summary: List household members
      parameters:
        - $ref: "#/components/parameters/HouseholdId"
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/HouseholdMember"
        "404":
          $ref: "#/components/responses/NotFound"
  /households/{householdId}/members/{memberId}:
    parameters:
      - $ref: "#/components/parameters/HouseholdId"
      - $ref: "#/components/parameters/MemberId"
    patch:
      tags: [Members]
      operationId: updateHouseholdMember
      summary: Update a member role
      description: Owner only. The last owner cannot be demoted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  $ref: "#/components/schemas/HouseholdRole"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdMember"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Members]
      operationId: deleteHouseholdMember
      summary: Remove a member or leave a household
      responses:
        "204":
          description: Membership removed
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /households/{householdId}/invitations:
    get:
      tags: [Invitations]
      operationId: listHouseholdInvitations
      summary: List active household invitations
      description: Owner only.
      parameters:
        - $ref: "#/components/parameters/HouseholdId"
      responses:
        "200":
          description: Invitation list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/InvitationSummary"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Invitations]
      operationId: createHouseholdInvitation
      summary: Create a household invitation
      description: Owner only. The raw token is returned once.
      parameters:
        - $ref: "#/components/parameters/HouseholdId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedInvitation"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /households/{householdId}/invitations/{invitationId}:
    delete:
      tags: [Invitations]
      operationId: revokeHouseholdInvitation
      summary: Revoke an invitation
      description: Owner only.
      parameters:
        - $ref: "#/components/parameters/HouseholdId"
        - $ref: "#/components/parameters/InvitationId"
      responses:
        "204":
          description: Invitation revoked
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /invitations/{token}:
    get:
      tags: [Invitations]
      operationId: previewInvitation
      security: []
      summary: Preview an invitation from a deep link
      parameters:
        - $ref: "#/components/parameters/InvitationToken"
      responses:
        "200":
          description: Safe invitation preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationPreview"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
  /invitations/{token}/accept:
    post:
      tags: [Invitations]
      operationId: acceptInvitation
      summary: Accept an invitation
      parameters:
        - $ref: "#/components/parameters/InvitationToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /households/{householdId}/rooms:
    get:
      tags: [Rooms]
      operationId: listRooms
      summary: List household rooms
      parameters:
        - $ref: "#/components/parameters/HouseholdId"
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Room list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoomSummary"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Rooms]
      operationId: createRoom
      summary: Create a room
      parameters:
        - $ref: "#/components/parameters/HouseholdId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /rooms/{roomId}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [Rooms]
      operationId: getRoom
      summary: Get room details
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Rooms]
      operationId: updateRoom
      summary: Update, reorder, archive, or restore a room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomRequest"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Rooms]
      operationId: archiveRoom
      summary: Archive a room
      parameters:
        - $ref: "#/components/parameters/ExpectedVersion"
      responses:
        "204":
          description: Room archived
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /households/{householdId}/plans/generate:
    post:
      tags: [Plans]
      operationId: generateDailyPlan
      summary: Create or regenerate a daily plan
      parameters:
        - $ref: "#/components/parameters/HouseholdId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePlanRequest"
      responses:
        "200":
          description: Existing plan regenerated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /households/{householdId}/daily-plans/{localDate}:
    get:
      tags: [Plans]
      operationId: getDailyPlanByDate
      summary: Get the current user's plan for a local date
      parameters:
        - $ref: "#/components/parameters/HouseholdId"
        - $ref: "#/components/parameters/LocalDate"
      responses:
        "200":
          description: Daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "404":
          $ref: "#/components/responses/NotFound"
  /plans/{planId}:
    parameters:
      - $ref: "#/components/parameters/PlanId"
    get:
      tags: [Plans]
      operationId: getDailyPlan
      summary: Get a daily plan
      responses:
        "200":
          description: Daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Plans]
      operationId: updateDailyPlan
      summary: Update status, budget, energy, or scheduled calendar block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanRequest"
      responses:
        "200":
          description: Updated daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /plans/{planId}/items:
    post:
      tags: [Plans]
      operationId: addDailyPlanItem
      summary: Add a task to a plan manually
      parameters:
        - $ref: "#/components/parameters/PlanId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPlanItemRequest"
      responses:
        "200":
          description: Updated daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /plans/{planId}/items/{itemId}:
    parameters:
      - $ref: "#/components/parameters/PlanId"
      - $ref: "#/components/parameters/PlanItemId"
    patch:
      tags: [Plans]
      operationId: updateDailyPlanItem
      summary: Skip, defer, restore, or reorder a plan item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanItemRequest"
      responses:
        "200":
          description: Updated daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Plans]
      operationId: deleteDailyPlanItem
      summary: Remove an item from a plan
      parameters:
        - $ref: "#/components/parameters/ExpectedVersion"
      responses:
        "200":
          description: Updated daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token.
  parameters:
    HouseholdId:
      name: householdId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InvitationId:
      name: invitationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InvitationToken:
      name: token
      in: path
      required: true
      schema:
        type: string
        minLength: 32
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PlanId:
      name: planId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PlanItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LocalDate:
      name: localDate
      in: path
      required: true
      schema:
        type: string
        format: date
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
    ExpectedVersion:
      name: expectedVersion
      in: query
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
    Unauthorized:
      description: Authentication is required or the token is invalid
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: The authenticated user does not have permission
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found within the authorization boundary
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Domain or resource version conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Gone:
      description: Resource has expired or was revoked
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds before retrying.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    HouseholdRole:
      type: string
      enum: [owner, member]
    HouseholdSummary:
      type: object
      required: [id, name, timezone, role, memberCount, roomCount, dueTaskCount, version]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
        role:
          $ref: "#/components/schemas/HouseholdRole"
        memberCount:
          type: integer
          minimum: 1
        roomCount:
          type: integer
          minimum: 0
        dueTaskCount:
          type: integer
          minimum: 0
        version:
          type: integer
          minimum: 1
    Household:
      allOf:
        - $ref: "#/components/schemas/HouseholdSummary"
        - type: object
          required: [createdBy, createdAt, updatedAt]
          properties:
            createdBy:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    HouseholdMember:
      type: object
      required: [id, userId, displayName, role, joinedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        avatarUrl:
          type: [string, "null"]
          format: uri
        role:
          $ref: "#/components/schemas/HouseholdRole"
        joinedAt:
          type: string
          format: date-time
    CreateInvitationRequest:
      type: object
      additionalProperties: false
      properties:
        expiresInHours:
          type: integer
          minimum: 1
          maximum: 168
          default: 72
    InvitationSummary:
      type: object
      required: [id, householdId, role, expiresAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/HouseholdRole"
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreatedInvitation:
      allOf:
        - $ref: "#/components/schemas/InvitationSummary"
        - type: object
          required: [token, deepLink]
          properties:
            token:
              type: string
            deepLink:
              type: string
              format: uri
    InvitationPreview:
      type: object
      required: [householdName, invitedByName, role, expiresAt]
      properties:
        householdName:
          type: string
        invitedByName:
          type: string
        role:
          $ref: "#/components/schemas/HouseholdRole"
        expiresAt:
          type: string
          format: date-time
    RoomType:
      type: string
      enum: [kitchen, bathroom, bedroom, living_room, hallway, office, other]
    RoomSummary:
      type: object
      required: [id, householdId, name, type, position, dueTaskCount, taskCount, archived, version]
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: "#/components/schemas/RoomType"
        position:
          type: integer
          minimum: 0
        dueTaskCount:
          type: integer
          minimum: 0
        taskCount:
          type: integer
          minimum: 0
        archived:
          type: boolean
        version:
          type: integer
          minimum: 1
    Room:
      allOf:
        - $ref: "#/components/schemas/RoomSummary"
        - type: object
          required: [createdAt, updatedAt]
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CreateRoomRequest:
      type: object
      additionalProperties: false
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        type:
          $ref: "#/components/schemas/RoomType"
        position:
          type: integer
          minimum: 0
    UpdateRoomRequest:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 80
        type:
          $ref: "#/components/schemas/RoomType"
        position:
          type: integer
          minimum: 0
        archived:
          type: boolean
...
    EnergyLevel:
      type: string
      enum: [low, normal, high]
    PlanStatus:
      type: string
      enum: [draft, active, completed, cancelled]
    PlanItemStatus:
      type: string
      enum: [planned, completed, skipped, deferred]
    SelectionReason:
      type: string
      enum: [overdue, due, priority, manual]
    DailyPlanItem:
      type: object
      required: [id, task, estimatedMinutes, selectionReason, position, status]
      properties:
        id:
          type: string
          format: uuid
        task:
          $ref: "#/components/schemas/TaskSummary"
        assignedTo:
          type: [string, "null"]
          format: uuid
        estimatedMinutes:
          type: integer
          minimum: 1
        selectionReason:
          $ref: "#/components/schemas/SelectionReason"
        position:
          type: integer
          minimum: 0
        status:
          $ref: "#/components/schemas/PlanItemStatus"
        completionId:
          type: [string, "null"]
          format: uuid
    DailyPlan:
      type: object
      required:
        - id
        - householdId
        - userId
        - localDate
        - timezone
        - timeBudgetMinutes
        - energyLevel
        - status
        - items
        - totalEstimatedMinutes
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        localDate:
          type: string
          format: date
        timezone:
          type: string
        timeBudgetMinutes:
          type: integer
          minimum: 5
          maximum: 240
        energyLevel:
          $ref: "#/components/schemas/EnergyLevel"
        status:
          $ref: "#/components/schemas/PlanStatus"
        scheduledStartAt:
          type: [string, "null"]
          format: date-time
        scheduledEndAt:
          type: [string, "null"]
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/DailyPlanItem"
        totalEstimatedMinutes:
          type: integer
          minimum: 0
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GeneratePlanRequest:
      type: object
      additionalProperties: false
      required: [localDate, timezone, timeBudgetMinutes, energyLevel]
      properties:
        localDate:
          type: string
          format: date
        timezone:
          type: string
        timeBudgetMinutes:
          type: integer
          minimum: 5
          maximum: 240
        energyLevel:
          $ref: "#/components/schemas/EnergyLevel"
        roomIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        includeAssignedToOthers:
          type: boolean
          default: false
        preserveManualItems:
          type: boolean
          default: true
        expectedPlanVersion:
          type: [integer, "null"]
          minimum: 1
    UpdatePlanRequest:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        timeBudgetMinutes:
          type: integer
          minimum: 5
          maximum: 240
        energyLevel:
          $ref: "#/components/schemas/EnergyLevel"
        status:
          $ref: "#/components/schemas/PlanStatus"
        scheduledStartAt:
          type: [string, "null"]
          format: date-time
        scheduledEndAt:
          type: [string, "null"]
          format: date-time
    AddPlanItemRequest:
      type: object
      additionalProperties: false
      required: [planVersion, taskId]
      properties:
        planVersion:
          type: integer
          minimum: 1
        taskId:
          type: string
          format: uuid
        position:
          type: integer
          minimum: 0
    UpdatePlanItemRequest:
      type: object
      additionalProperties: false
      required: [planVersion]
      properties:
        planVersion:
          type: integer
          minimum: 1
        status:
          $ref: "#/components/schemas/PlanItemStatus"
        position:
          type: integer
          minimum: 0
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
        detail:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
        currentResource:
          type: [object, "null"]
          additionalProperties: true